+++ android-core/src/androidTest/kotlin/com.mparticle/networking/NetworkOptionsTest.kt
@@ -3,6 +3,7 @@ package com.mparticle.networking
 import com.mparticle.MParticle
 import com.mparticle.MParticleOptions
 import com.mparticle.internal.AccessUtils
+import com.mparticle.internal.database.UploadSettings
 import com.mparticle.testutils.BaseCleanInstallEachTest
 import org.junit.After
 import org.junit.Assert
@@ -22,7 +23,7 @@ class NetworkOptionsTest : BaseCleanInstallEachTest() {
         startMParticle(MParticleOptions.builder(mContext).credentials(apiKey, "secret"))
         setClients()
         for (endpoint in MParticleBaseClientImpl.Endpoint.values()) {
-            defaultUrls[endpoint] = mpClient.getUrl(endpoint, endpoint.name, null)
+            defaultUrls[endpoint] = mpClient.getUrl(endpoint, endpoint.name, null,UploadSettings(apiKey, "secret", NetworkOptions.builder().build(), "", ""))
         }
         MParticle.setInstance(null)
     }
@@ -240,7 +241,7 @@ class NetworkOptionsTest : BaseCleanInstallEachTest() {
         var randIdentityPath = mRandomUtils.getAlphaString(10)
         Assert.assertEquals(
             "/v1/$randIdentityPath",
-            mpClient.getUrl(MParticleBaseClientImpl.Endpoint.IDENTITY, randIdentityPath).path
+            mpClient.getUrl(MParticleBaseClientImpl.Endpoint.IDENTITY, randIdentityPath, null).path
         )
         Assert.assertEquals(
             audienceUrl,
@@ -263,7 +264,8 @@ class NetworkOptionsTest : BaseCleanInstallEachTest() {
             "/v1/$randIdentityPath",
             identityClient.getUrl(
                 MParticleBaseClientImpl.Endpoint.IDENTITY,
-                randIdentityPath
+                randIdentityPath,
+                null
             ).path
         )
 
@@ -284,7 +286,8 @@ class NetworkOptionsTest : BaseCleanInstallEachTest() {
             defaultUrls[MParticleBaseClientImpl.Endpoint.IDENTITY]?.path,
             mpClient.getUrl(
                 MParticleBaseClientImpl.Endpoint.IDENTITY,
-                MParticleBaseClientImpl.Endpoint.IDENTITY.name
+                MParticleBaseClientImpl.Endpoint.IDENTITY.name,
+                null
             ).path
         )
     }