+++ android-core/src/main/java/com/mparticle/internal/ConfigManager.java
@@ -4,6 +4,7 @@ import android.content.Context;
 import android.content.SharedPreferences;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
+import android.net.Network;
 import android.os.Build;
 
 import androidx.annotation.NonNull;
@@ -16,6 +17,7 @@ import com.mparticle.MParticle;
 import com.mparticle.MParticleOptions;
 import com.mparticle.consent.ConsentState;
 import com.mparticle.identity.IdentityApi;
+import com.mparticle.internal.database.UploadSettings;
 import com.mparticle.internal.messages.BaseMPMessage;
 import com.mparticle.networking.NetworkOptions;
 import com.mparticle.networking.NetworkOptionsManager;
@@ -338,7 +340,7 @@ public class ConfigManager {
         }
     }
 
-    void clearConfig() {
+    public void clearConfig() {
         sPreferences.edit()
                 .remove(CONFIG_JSON)
                 .remove(CONFIG_JSON_TIMESTAMP)
@@ -511,6 +513,26 @@ public class ConfigManager {
         }
     }
 
+    public String getSupportedKitString() {
+        MParticle instance = MParticle.getInstance();
+        if (instance != null) {
+            Set<Integer> supportedKitIds = instance.Internal().getKitManager().getSupportedKits();
+            if (supportedKitIds != null && !supportedKitIds.isEmpty()) {
+                StringBuilder buffer = new StringBuilder(supportedKitIds.size() * 3);
+                Iterator<Integer> it = supportedKitIds.iterator();
+                while (it.hasNext()) {
+                    Integer next = it.next();
+                    buffer.append(next);
+                    if (it.hasNext()) {
+                        buffer.append(",");
+                    }
+                }
+                return buffer.toString();
+            }
+        }
+        return "";
+    }
+
     /**
      * When the Config manager starts up, we don't want to enable everything immediately to save on app-load time.
      * This method will be called from a background thread after startup is already complete.
@@ -574,6 +596,16 @@ public class ConfigManager {
 
     public void setLogUnhandledExceptions(boolean log) {
         sPreferences.edit().putBoolean(Constants.PrefKeys.REPORT_UNCAUGHT_EXCEPTIONS, log).apply();
+    }
+
+    public UploadSettings getUploadSettings() {
+        String apiKey = getApiKey();
+        String secret = getApiSecret();
+        if (apiKey == null || secret == null) {
+            return null;
+        }
+
+        return new UploadSettings(apiKey, secret, getNetworkOptions(), getActiveModuleIds(), getSupportedKitString());
     }
 
     public String getApiKey() {
@@ -1299,10 +1331,10 @@ public class ConfigManager {
      silo and a hyphen (ex. "us1-", "us2-", "eu1-").  us1 was the first silo,and before other silos existed, there were no prefixes and all apiKeys
      were us1. As such, if we split on a '-' and the resulting array length is 1, then it is an older APIkey that should route to us1.
      When splitKey.length is greater than 1, then splitKey[0] will be us1, us2, eu1, au1, or st1, etc as new silos are added */
-    public String getPodPrefix() {
+    public String getPodPrefix(@NonNull String apiKey) {
         String prefix = "us1";
         try {
-            String[] prefixFromApi = getApiKey().split("-");
+            String[] prefixFromApi = apiKey.split("-");
             if (prefixFromApi.length > 1) {
                 prefix = prefixFromApi[0];
             }
@@ -1331,6 +1363,14 @@ public class ConfigManager {
     public synchronized void setNetworkOptions(NetworkOptions networkOptions) {
         sNetworkOptions = networkOptions;
         sPreferences.edit().remove(Constants.PrefKeys.NETWORK_OPTIONS).apply();
+    }
+
+    public UploadSettings getLastUploadSettings() {
+        return getUserStorage().getLastUploadSettings();
+    }
+
+    public void setLastUploadSettings(@NonNull UploadSettings uploadSettings) {
+        getUserStorage().setLastUploadSettings(uploadSettings);
     }
 
     @NonNull