+++ android-core/src/main/java/com/mparticle/internal/database/services/SQLiteOpenHelperWrapper.java
@@ -9,5 +9,4 @@ public interface SQLiteOpenHelperWrapper {
     void onUpgrade(SQLiteDatabase database, int oldVersion, int newVersion);
 
     void onDowngrade(SQLiteDatabase database, int oldVersion, int newVersion);
-
 }