+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+## [5.59.0](https://github.com/mParticle/mparticle-android-sdk/compare/v5.58.12...v5.59.0) (2025-04-03)
+
+
+### Features
+
+* Add Workspace Switching feature ([#551](https://github.com/mParticle/mparticle-android-sdk/issues/551)) ([de72f0a](https://github.com/mParticle/mparticle-android-sdk/commit/de72f0a0afe0917e1550ea9a31e212a9682fc4f4))
+
+
+### Updates & Maintenance
+
+* Downgrade API level to 28 for both instrument test and instrument… ([#552](https://github.com/mParticle/mparticle-android-sdk/issues/552)) ([ddf8f4f](https://github.com/mParticle/mparticle-android-sdk/commit/ddf8f4f19dd159193e9c80a764270f4eadc1973f))
+* Downgrade API level to 28 for daily yml file ([#553](https://github.com/mParticle/mparticle-android-sdk/issues/553)) ([12eef85](https://github.com/mParticle/mparticle-android-sdk/commit/12eef852fe8329d46631206f158ec5c77bbacf2c))
+
+## [5.58.12](https://github.com/mParticle/mparticle-android-sdk/compare/v5.58.11...v5.58.12) (2025-03-18)
+
+
+### Updates & Maintenance
+
+* Update submodules ([01c5f37](https://github.com/mParticle/mparticle-android-sdk/commit/01c5f37367fc928f8e3c42105f8af5b755bfb113))
+
+## [5.58.11](https://github.com/mParticle/mparticle-android-sdk/compare/v5.58.10...v5.58.11) (2025-03-05)
+
+
+### Bug Fixes
+
+* Update Kotlin version to be compatible with Braze ([#547](https://github.com/mParticle/mparticle-android-sdk/issues/547)) ([50d73b3](https://github.com/mParticle/mparticle-android-sdk/commit/50d73b3168db250626afe6641bc691173c2024f4))
+
+
+### Updates & Maintenance
+
+* Update submodules ([b9ccc6c](https://github.com/mParticle/mparticle-android-sdk/commit/b9ccc6c68001267d0a962fdcbf9f58298f01ce0f))
+
+## [5.58.10](https://github.com/mParticle/mparticle-android-sdk/compare/v5.58.9...v5.58.10) (2025-02-27)
+
+
+### Bug Fixes
+
+* previous Google ad Id from being null on cold launch ([#546](https://github.com/mParticle/mparticle-android-sdk/issues/546)) ([3051a82](https://github.com/mParticle/mparticle-android-sdk/commit/3051a823c48720c4810ace37f8a5241768ede0fc))
+
+
+### Updates & Maintenance
+
+* Migrate Internal Application context class to kotlin ([#545](https://github.com/mParticle/mparticle-android-sdk/issues/545)) ([ecd2833](https://github.com/mParticle/mparticle-android-sdk/commit/ecd2833a505059b95654b684b771595c92a146eb))
+* Update submodules ([e9cb498](https://github.com/mParticle/mparticle-android-sdk/commit/e9cb49839ad7c77beff5ef8e6d3db60073b7ec21))
+
 ## [5.58.9](https://github.com/mParticle/mparticle-android-sdk/compare/v5.58.8...v5.58.9) (2025-02-13)